backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic-settings
redis
celery
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
twilio
firebase-admin
boto3
geoalchemy2
httpx
python-dotenv
